[EAGLE-5] Merge general-purpose monitor app into eagle-core
