+++ eagle-core/eagle-data-process/eagle-stream-process-api/src/main/java/eagle/dataproc/impl/storm/kafka/JsonMessageDeserializer.scala
@@ -14,14 +14,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eagle.monitor
+package eagle.dataproc.impl.storm.kafka
 
-import eagle.datastream.StormStreamApp
+import java.io.IOException
+import java.util
+import java.util.Properties
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import org.slf4j.{Logger, LoggerFactory}
 
 /**
  * @since  11/6/15
  */
-abstract class StreamMonitor extends StormStreamApp{
-  implicit val streamName = get[String]("eagle.stream.name","eventStream")
-  implicit val streamExecutorId = get[String]("eagle.stream.executor",s"${streamName}Executor")
+case class JsonMessageDeserializer(props:Properties) extends SpoutKafkaMessageDeserializer{
+  private val objectMapper: ObjectMapper = new ObjectMapper
+  private val LOG: Logger = LoggerFactory.getLogger(classOf[JsonMessageDeserializer])
+
+  override def deserialize(bytes: Array[Byte]): AnyRef = {
+    var map: util.Map[String, _] = null
+    try {
+      map = objectMapper.readValue(bytes, classOf[util.TreeMap[String, _]])
+    } catch {
+      case e: IOException => {
+        LOG.error("Failed to deserialize json from: " + new String(bytes), e)
+      }
+    }
+    map
+  }
 }
