+++ eagle-core/eagle-data-process/eagle-stream-process-api/src/test/java/eagle/datastream/TestKafkaStreamMonitor.java
@@ -1,0 +1,13 @@
+package eagle.datastream;
+
+import eagle.dataproc.impl.storm.kafka.KafkaStreamMonitorApp;
+
+/**
+ * @author Chen, Hao (hchen9@ebay.com)
+ * @since 11/7/15
+ */
+public class TestKafkaStreamMonitor {
+    public static void main(String[] args){
+        new KafkaStreamMonitorApp().main(args);
+    }
+}
