+++ eagle-core/eagle-data-process/eagle-stream-process-api/src/test/resources/application.conf
@@ -1,35 +1,48 @@
 {
-  "envContextConfig" : {
-    "env" : "storm",
-    "mode" : "local",
-    "topologyName" : "testTopology",
-    "parallelismConfig" : {
-      "prepend" : 3,
-      "testSpout1" : 2
-    }
-  },
-  "eagleProps" : {
-    "site" : "sandbox",
-    "dataSource": "ds1",
-    "eagleService" : {
-      "host": "localhost",
-      "port": 38080
-    },
-    "dataJoinPollIntervalSec" : 30,
-    "mailHost" : "mailhost.com",
-    "mailSmtpPort":"25",
-    "mailDebug" : "true"
-  },
-  "alertExecutorConfigs" : {
-    "alert1" : {
-      "parallelism" : 2,
-      "partitioner" : "eagle.alert.policy.DefaultPolicyPartitioner"
-      "needValidation" : "true"
-    }
-  },
-  "dynamicConfigSource" : {
-    "enabled" : true,
-    "initDelayMillis" : 0,
-    "delayMillis" : 30000
-  }
+	"envContextConfig" : {
+		"env" : "storm",
+		"mode" : "local",
+		"topologyName" : "kafka-monitor-topology",
+		"parallelismConfig" : {
+			"kafkaMsgConsumer" : 1
+		}
+	},
+	"dataSourceConfig": {
+		"topic" : "sample_event",
+		"zkConnection" : "localhost:2181",
+		"zkConnectionTimeoutMS" : 15000,
+		"consumerGroupId" : "EagleConsumer",
+		"fetchSize" : 1048586,
+		"deserializerClass" : "eagle.dataproc.impl.storm.kafka.JsonMessageDeserializer",
+		"transactionZKServers" : "localhost",
+		"transactionZKPort" : 2181,
+		"transactionZKRoot" : "/consumers",
+		"transactionStateUpdateMS" : 2000
+	},
+	"alertExecutorConfigs" : {
+		"eventStreamExecutor" : {
+			"parallelism" : 1,
+			"partitioner" : "eagle.alert.policy.DefaultPolicyPartitioner"
+			"needValidation" : "true"
+		}
+	},
+	"eagleProps" : {
+		"site" : "sandbox",
+		"dataSource": "eventSource",
+		"dataJoinPollIntervalSec" : 30,
+		"mailHost" : "mail.host.com",
+		"mailSmtpPort":"25",
+		"mailDebug" : "true",
+		"eagleService": {
+			"host": "localhost",
+			"port": 38080,
+			"username": "admin",
+			"password": "secret"
+		}
+	},
+	"dynamicConfigSource" : {
+		"enabled" : true,
+		"initDelayMillis" : 0,
+		"delayMillis" : 30000
+	}
 }
