+++ eagle-core/eagle-data-process/eagle-stream-process-api/src/test/resources/event-metadata-init.sh
