--- eagle-experimental/eagle-monitor/src/main/resources/application.conf
@@ -1,49 +1,0 @@
-{
-	"envContextConfig" : {
-		"env" : "storm",
-		"mode" : "local",
-		"topologyName" : "auditLogProcessTopology",
-		"parallelismConfig" : {
-			"kafkaMsgConsumer" : 1,
-			"hdfsAuditLogAlertExecutor*" : 1
-		}
-	},
-	"dataSourceConfig": {
-		"topic" : "sample_event",
-		"zkConnection" : "sandbox.hortonworks.com:2181",
-		"zkConnectionTimeoutMS" : 15000,
-		"consumerGroupId" : "EagleConsumer",
-		"fetchSize" : 1048586,
-		"deserializerClass" : "eagle.monitor.JsonDeserializer",
-		"transactionZKServers" : "localhost",
-		"transactionZKPort" : 2181,
-		"transactionZKRoot" : "/consumers",
-		"transactionStateUpdateMS" : 2000
-	},
-	"alertExecutorConfigs" : {
-		"eventStreamExecutor" : {
-			"parallelism" : 1,
-			"partitioner" : "eagle.alert.policy.DefaultPolicyPartitioner"
-			"needValidation" : "true"
-		}
-	},
-	"eagleProps" : {
-		"site" : "sandbox",
-		"dataSource": "eventSource",
-		"dataJoinPollIntervalSec" : 30,
-		"mailHost" : "mail.host.com",
-		"mailSmtpPort":"25",
-		"mailDebug" : "true",
-		"eagleService": {
-			"host": "localhost",
-			"port": 38080,
-			"username": "admin",
-			"password": "secret"
-		}
-	},
-	"dynamicConfigSource" : {
-		"enabled" : true,
-		"initDelayMillis" : 0,
-		"delayMillis" : 30000
-	}
-}
