--- eagle-experimental/eagle-monitor/src/main/scala/eagle/monitor/KafkaStreamMonitor.scala
@@ -1,29 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package eagle.monitor
-
-import eagle.dataproc.impl.storm.kafka.KafkaSourcedSpoutProvider
-
-/**
- * @since  11/6/15
- */
-object KafkaStreamMonitor extends StreamMonitor{
-  set("dataSourceConfig.deserializerClass",classOf[JsonDeserializer])
-
-  source(new KafkaSourcedSpoutProvider).renameOutputFields(1).withName(streamName)
-    .alertWithConsumer(streamName, streamExecutorId)
-}
